+++ source/config/default.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
  */
 return [
     'app' => [
-        'version' => '0.3.0.2',
+        'version' => '0.3.0.3',
     ],
     'site' => [
         'name' => 'Latch',